refactor(handlers): abstract update delay script logic
